refactor the functions to be used in drag and insert
